+++ src/util.js
@@ -23,9 +23,9 @@ define(function (require) {
      *
      * 传入一个字符串时，只返回一个发送器函数；传入数组或对象时，递归；传入函数时
      *
-     * @param {string|Array.<string>|Object.<string, string>|function} url 请求路径或多个请求路径的集合，或是取值函数
+     * @param {string|Array.<string>|Object.<string, string>|Function} url 请求路径或多个请求路径的集合，或是取值函数
      * @param {function(string):boolean} isRequester 判断是否是需要生成请求发送器的路径
-     * @return {function|Array.<function>|Object.<string, function>} 将对应的路径转换为发送器后返回
+     * @return {Function|Array.<Function>|Object.<string, Function>} 将对应的路径转换为发送器后返回
      */
     util.genRequesters = function (url, isRequester) {
         if (u.typeOf(url) === 'String') {
@@ -328,7 +328,45 @@ define(function (require) {
     };
 
     /**
+     * 生成列表页项，可选带省略，带TipLayer
+     *
+     * @param  {Object} options                 可选参数配置
+     * @param  {string} [options.className]     类
+     * @param  {string} [options.title]         交给QuickTip的title
+     * @param  {string} [options.content]       最终展示的内容，考虑到可能还要加其他东西，
+     *                                          自己在调用前对后端的内容escape一下吧
+     * @param  {boolean} [options.isEllipsis]   是否超长省略
+     * @return {string}                         html
+     */
+    util.genTruncatableContent = function (options) {
+        var content = options.content || '';
+        var title = options.title || '';
+        var isEllipsis = options.isEllipsis || false;
+        var className = options.className || '';
+
+        var tpl = '<span';
+        if (isEllipsis && className) {
+            tpl += ' class="list-ellipsis ' + className + '"';
+        }
+        else if (isEllipsis) {
+            tpl += ' class="list-ellipsis"';
+        }
+        else if (className) {
+            tpl += ' class="' + className + '"';
+        }
+
+        if (title) {
+            tpl += ' data-title="' + u.escape(title) + '"';
+        }
+
+        tpl += '>' + content + '</span>';
+
+        return tpl;
+    };
+
+    /**
      * 下载文件
+     *
      * @param {string} url 文件地址.
      */
     util.download = function (url) {