+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+* 0.2.6
+    - 增加 `system/session` 模块
+    - `ListView` 原型增加 `batchModify` 方法，提供批量操作在 View 层的扩展点
+    - 增加表单元素的字体默认设置
+    - `tpl` 模块支持解析 `data-ui` 属性中以集中方式设置的 `type`
+    - 修复 `.list-operation` 样式在 Firefox 下的问题
+    - 调整了图表控件中 `echarts` 的 `formatter` 参数
+    - 增加了 `genTruncatableContent` 来在列表中生成超长截断的文本，并为 `QuickTip` 扩展提供对应全部文本作为参数
+    - 修正代码规范
+
 * 0.2.5
     - `QuickTip` 改用 `TipLayer` 的 `manual` 模式
     - `RichBoxGroup` 支持设置 `singleSelect`